+++ client/layouts/Main.jsx
@@ -1,15 +1,3 @@
-/*MainLayout = React.createClass({
-  render() {
-    return (
-      <div>
-        <Header />
-        <main>{this.props.content}</main>
-      </div>
-    );
-  }
-});
-*/
-
 import React from 'react';
 import ReactDOM from 'react-dom';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
@@ -19,6 +7,7 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 MainLayout = React.createClass({
 
   checkSize() {
+    // Minimise menu when screen is less than 768px
     if (window.matchMedia('(max-width: 768px)').matches) {
       $('body').addClass('body-small fixed-sidebar')
     } else {
@@ -45,40 +34,15 @@ MainLayout = React.createClass({
   },
 
   componentDidMount() {
-    // Minimalize menu when screen is less than 768px
     this.checkSize();
     $(window).bind("resize", this.checkSize);
-
-    // SKIN OPTIONS
-    // Uncomment this if you want to have different skin option:
-    // Available skin: (skin-1 or skin-3, skin-2 deprecated, md-skin)
      $('body').addClass('md-skin');
-
-    // FIXED-SIDEBAR
-    // Uncomment this if you want to have fixed left navigation
-    /*
-    $('body').addClass('fixed-sidebar');
-     $('.sidebar-collapse').slimScroll({
-         height: '100%',
-         railOpacity: 0.9
-     });
-    */
-    
-    // slim scroll for right sidebar
-    $('.sidebar-container').slimScroll({
-        height: '100%',
-        railOpacity: 0.4,
-        wheelStep: 10
-    });
-
-    // BOXED LAYOUT
-    // Uncomment this if you want to have boxed layout
-    // $('body').addClass('boxed-layout');
-
   },
+
   componentWillUnmount() {
      $('body').removeClass('md-skin');
   },
+
   render() {
     return (
       <MuiThemeProvider muiTheme={getMuiTheme()}>
@@ -98,4 +62,3 @@ MainLayout = React.createClass({
     );
   }
 });
-